feat: added limiter and refactored cors
