+++ app.js
@@ -4,21 +4,24 @@ const express = require('express');
 const mongoose = require('mongoose');
 const cookieParser = require('cookie-parser');
 const helmet = require('helmet');
+const cors = require('cors');
 const { errors } = require('celebrate');
 
 const { PORT = 3000 } = process.env;
 
 const router = require('./routes/index');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
-const { mongooseConnectURI, mongooseConnectOptions } = require('./constants/constants');
+const limiter = require('./middlewares/limiter');
+
+const { mongooseConnectURI, mongooseConnectOptions } = require('./utils/constants');
 const errorHandler = require('./middlewares/error-handler');
-const corsHandler = require('./middlewares/cors-handler');
 
 const app = express();
 
 app.use(helmet());
 
 app.use(requestLogger);
+app.use(limiter);
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
@@ -27,7 +30,12 @@ app.use(cookieParser());
 
 mongoose.connect(mongooseConnectURI, mongooseConnectOptions);
 
-app.use(corsHandler);
+app.use(cors({
+  origin: [
+    'localhost:3000',
+  ],
+  credentials: true,
+}));
 
 app.use(router);
