--- middlewares/cors-handler.js
@@ -1,18 +1,0 @@
-const { allowedCors, DEFAULT_ALLOWED_METHODS } = require('../constants/constants');
-
-const corsHandler = (req, res, next) => {
-  const { origin } = req.headers;
-  const { method } = req;
-  const requestHeaders = req.headers['access-control-request-headers'];
-
-  if (allowedCors.includes(origin)) {
-    res.header('Access-Control-Allow-Origin', origin);
-  }
-  if (method === 'OPTIONS') {
-    res.header('Access-Control-Allow-Methods', DEFAULT_ALLOWED_METHODS);
-    res.header('Access-Control-Allow-Headers', requestHeaders);
-  }
-  next();
-};
-
-module.exports = corsHandler;
