+++ utils/constants.js
@@ -7,15 +7,7 @@ const mongooseConnectOptions = {
   useUnifiedTopology: true,
 };
 
-const allowedCors = [
-  'localhost:3000',
-];
-
-const DEFAULT_ALLOWED_METHODS = 'GET,HEAD,PUT,PATCH,POST,DELETE';
-
 module.exports = {
   mongooseConnectURI,
   mongooseConnectOptions,
-  allowedCors,
-  DEFAULT_ALLOWED_METHODS,
 };
